added test/SigDefBug.trellys illustrating a (probaby known) bug, and modified the test/Makefile to be quiet on regression testing and have a separate TODO rule.
I have an idea how to fix the the SigDefBug bug, and will start on that next.  I realize this is not an urgent bug, but it will give me chance to learn more about the typechecker code.

I added the TODO rule, instead adding the TODO items to the existing PASS and FAIL lists, so that you don't have keep track of what worked before you edited: make all will always be silently if you didn't break any previouvly correct test cases.

git-svn-id: https://trellys.googlecode.com/svn/trunk@39 59285afe-504e-ee2a-6240-25eb0d263adb