Sandboxed the build/install procedure and updated/expanded the documentation.
NB: the new setup still requires that you have the default Cabal
install directory on your $PATH.  That directory is ~/.cabal/bin for
the default "per-user" Cabal setup, which you are probably using.  It
needs to be on your path, because the new setup `cabal install`s
Capri there.

The old setup put the trellys binary in the default Cabal install dir.
This is no longer the case.  The trellys binary is now symlinked to
test/trellys, the only place you're expected to be running `trellys -t`.


git-svn-id: https://trellys.googlecode.com/svn/trunk@51 59285afe-504e-ee2a-6240-25eb0d263adb