Fixed issue 8: added "axiom" keyword to introduce typings into context w/o corresponding definitions.
Probably nothing surprising here, although I added a recursive (Axiom
Decl) constructor to the Decl type that is only supposed to be applied
to Sigs.  This made the other code I had to write minimal, but it
might be better style to create an AxeSig constructor that duplicates
Sig, but is interpreted differently (as an axiom).  Thoughts?


git-svn-id: https://trellys.googlecode.com/svn/trunk@52 59285afe-504e-ee2a-6240-25eb0d263adb