Support _ for unimportant variables and a->b for nondependent products (fixes issues 9 and 10).
The parser now tries to handle _ as a name in all binding positions, and
internaly creates a fresh new name for each occurance. _ does not quite behave
as a variable name,  since it is not supported in non-binding positions (I think
eventually we will want to parse those differently). Also, I could not figure
out how to make the Parsec grammar handle (_:a)->b, sorry! (You can write a->b
instead though).

Also, the syntax
  exp -> exp
  [exp] -> exp
  exp => exp
  [exp] => exp
is now supported as sugar for non-dependent products.

I've updated a few of the testcases to use the new syntax additions.



git-svn-id: https://trellys.googlecode.com/svn/trunk@61 59285afe-504e-ee2a-6240-25eb0d263adb