Fixed the free variable check for erased lambdas.
The fix is only two lines, but the bug is very subtle: the
introduction of ad-hoc polymorphism in a return type allowed a type
mismatch to result in calling the wrong version of the function,
instead of a type error.

Speculative details: in r71 Vilhelm integrated the new RepLib. This
included a switch from using Name for names in Terms and ETerms, to
using (Name Term) and (Name ETerm), resp.  In most cases where this
resulted in a type mismatch between (Name ETerm) and (Name Term), the
type checker complained and so Vilhelm inserted a translate call.
However, the free variables test uses the function fv, which is ad-hoc
polymorphic in it's return type, and so the associated (Name ETerm) vs
(Name Term) mismatch resulted in calling the wrong fv, instead of a
type error.

I find this very interesting: it's often suggested that static type
checking allows one to refactor confidently, unlike in a dynamically
checked setting, because the type checker will complain whereever the
refactoring is incomplete.  However, in the case of ad-hoc
polymorphism in the return type, a type mismatch introduced by
refactoring can result in calling the wrong function in a type correct
way, instead of a type error!

git-svn-id: https://trellys.googlecode.com/svn/trunk@108 59285afe-504e-ee2a-6240-25eb0d263adb