Fixed a bug with the way that imported modules are added to the context.
Previously, we were adding declarations from an imported module to the context
in the order they appeared in the imported module, rather than in the reverse
order. This caused substDefs to fail in the case where an imported function
depended on /another/ imported function (defined earlier in the module).

For example, if you imported Arithmetic, and then tried to typecheck

'join : mult 0 0 = 0,'

the typechecker will fail, becase (due to the improper order that imported
declarations are added to the context), 'plus' would be substituted *before* the
substitution of 'mult', and we'd get a stuck term in the evaluator where 'plus'
would be a free variable.

git-svn-id: https://trellys.googlecode.com/svn/trunk@115 59285afe-504e-ee2a-6240-25eb0d263adb