Pinned the external replib at revision 119 and added a pointer to our svn:externals doecs to the README.
When Vilhelm updated trellys to use the a newer replib in trellys r71 he removed the svn:externals replib revision pin instead of updating it.  This was not a problem as long as replib remained compatible with trellys, but the current replib is not compatible.  I pinnned replib at r119 because it's the first one I found that worked.  I have no idea if a newer replib revision would be better ...

git-svn-id: https://trellys.googlecode.com/svn/trunk@152 59285afe-504e-ee2a-6240-25eb0d263adb