Exhibit GHC 7.4.1 bug.
Commenting out unused code in one file causes a type error in another
file.  This looks like a GHC 7.4.1 bug.  Tim does not get the error
with his GHC (7.2?).  See the '!!!: GHC 7.4.1 bug ???' comments in
Monads.hs and Syntax.hs.

I need to isolate this and see if I should submit a bug report, but not right now ...

git-svn-id: https://trellys.googlecode.com/svn/trunk@540 59285afe-504e-ee2a-6240-25eb0d263adb